Moved Read map according to game phase
